--- a/pos-for-customer/order/pos-customer-order.page.ts
+++ b/pos-for-customer/order/pos-customer-order.page.ts
@@ -373,12 +373,6 @@ export class POSCustomerOrderPage extends PageBase {
             }
             else {
                 this.env.showPrompt('Bạn chắc muốn bỏ sản phẩm này khỏi giỏ hàng?', item.Name, 'Xóa sẩn phẩm').then(_ => {
-                    let tempQty = line.Quantity;
-                    tempQty += quantity;
-                    if (tempQty == 0 && this.item.OrderLines.length == 1) {
-                        this.env.showMessage('Đơn hàng phải có ít nhất 1 sản phẩm!','warning');
-                        return;
-                    }
                     line.Quantity += quantity;
                     this.calcOrderLine(line);
                     this.loadInfoOrder();