1. Update Printer Performance
Không dùng Promise all nữa, + điều chỉnh lại tham số để cải thiện.

Kết quả:
In bill / Receipt ( 1 phiếu ): 2s
In món ăn 1 vị trị ( 2 phiếu ) là 3s.
In món ăn 2 vị trí ( 2 phiếu ) là 4s.
In món ăn 2 vị trí ( 2 phiếu ) và có in food lẻ là 4s + 1s mỗi phiếu món ăn.

2. Chỉ tạo đơn mới khi nào bấm gửi bếp ( SendKitchen ).

3. Cho phép xóa toàn bộ món, sau đó casher / khách hàng chọn món mới.